requirements.txt
# 泵阀管道堵塞预警系统依赖

# 核心依赖
python>=3.9

# 数据处理库
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# AI模型组件
tensorflow>=2.10.0  # 或使用 torch>=2.0.0

# Web框架（模拟接口）
fastapi>=0.100.0
uvicorn>=0.23.0

# 其他依赖
pyyaml>=6.0
requests>=2.31.0
